pdf-gpt-summarizer/requirements.txt
streamlit
python-dotenv
PyPDF2
requests
